feat/DTFS2-4525 - split out TFM deal put api tests
