+++ dtfs-central-api/api-tests/v1/tfm/deal/tfm-gef-deal.api-test.js
@@ -39,40 +39,6 @@ describe('/v1/tfm/deal/:id', () => {
     });
   });
 
-  describe('PUT /v1/tfm/deal/:id', () => {
-    const dealUpdate = {
-      tfm: {
-        submissionDetails: {
-          exporterPartyUrn: '12345',
-        },
-      },
-    };
-
-    it('404s if updating an unknown id', async () => {
-      const { status } = await api.put({ dealUpdate }).to('/v1/tfm/deals/12345678');
-      expect(status).toEqual(404);
-    });
-
-    it('updates the created deal with correct fields, retaining original dealSnapshot', async () => {
-      const { body: portalDeal } = await api.post(newDeal).to('/v1/portal/gef/deals');
-      const dealId = portalDeal._id;
-
-      await api.put({
-        dealType: CONSTANTS.DEALS.DEAL_TYPE.GEF,
-        dealId,
-      }).to('/v1/tfm/deals/submit');
-
-      const { status, body } = await api.put({ dealUpdate }).to(`/v1/tfm/deals/${dealId}`);
-
-      expect(status).toEqual(200);
-      expect(body.dealSnapshot).toMatchObject(newDeal);
-      expect(body.tfm).toEqual({
-        ...dealUpdate.tfm,
-        lastUpdated: expect.any(Number),
-      });
-    });
-  });
-
   describe('PUT /v1/tfm/deal/:id/snapshot', () => {
     it('404s if updating an unknown id', async () => {
       const { status } = await api.put({}).to('/v1/tfm/deals/12345678/snapshot');
